pyproject.toml
[project]
name = "AIRWave"
version = "0.1.0"
description = "AI-driven information aggregation system for tracking academic and social trends"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "feedparser>=6.0.11",
    "anthropic>=0.39.0",
    "openai>=1.54.0",
    "google-genai>=0.3.0",
    "pydantic>=2.9.0",
    "rich>=13.9.0",
    "tenacity>=9.0.0",
    "python-dotenv>=1.0.0",
    "ddgs>=7.0.0",
    "beautifulsoup4>=4.12.0",
    "markdown>=3.10.2",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
airwave = "src.main:main"
airwave-mcp = "src.mcp.server:main"
airwave-wizard = "src.setup.wizard:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["tests"]
